LayoutTests/js/dfg-mul-big-integers-expected.txt

Tests what happens when you multiply large integers in an integer context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS foo(2147483647, 2147483646) is -2147483648
PASS successfullyParsed is true

TEST COMPLETE
